requirements.txt
black==24.3.0
flake8==6.1.0
pylint==3.0.2
pre-commit==3.7.0
requests==2.31.0
pandas==2.0.0
python-dotenv==1.0.1
pygsheets==2.0.6